backend-rest/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.4
requests==2.28.1
Pillow==9.3.0
Flask-Session==0.4.0
flask_cors==3.0.10